Resize Observer: Observe Device Pixel Content Box
In order to correctly snap box sizes to pixels using SnapSizeToPixel,
the function FragmentData::PaintOffset() needs to be used. This function
provides an accurate location/offset of the box in question. However,
this the paint offset is determined in the PrePaint lifecycle phase. To
accommodate this, LocalFrameView::NotifyResizeObservers is moved to
RunPrePaintLifecyclePhase from RunStyleAndLayoutLifecyclePhases.

Furthermore, contents of RunPrePaintLifecyclePhase are encapsulated into
a helper function. This helper function is then called by
NotifyResizeObservers after observations have been delivered and the
layout has been updated.

Resize Observer can now observe changes to the device pixel content box.
ResizeObserverEntry now reports the dimensions of the pixel snapped
content box.

Bug: 1055579
Change-Id: Ibc7b2e4e0a8468f9af74c8cfaadd02e4f0bcfc1f